Professionalized
Did the following:
closed all resources in finally blocks
checked whether delete calls actually worked before advancing
logged all errors in error text files
stopped methods after error is received
made Downloader constructors both call constructor method which reduces
total code in file
allowed property "user.dir" to be customizable
allowed URL to be pinged to test for internet connection to be
customizable
slightly improved readability with && instead of as many nested ifs
added comments (not very good, though!)
allowed downloads to be cancellable on the fly and the main file is not
lost

+++ src/gamepatcher/Downloader.java
@@ -13,11 +13,17 @@ import java.net.URLConnection;
 import java.net.UnknownHostException;
 import java.nio.channels.Channels;
 import java.nio.channels.ReadableByteChannel;
+import java.nio.file.DirectoryStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Calendar;
 import java.util.Scanner;
 import java.util.TimeZone;
 
-public class Downloader {
+import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
+
+public class Downloader extends Patcher {
 	private String filePath;
 	private String datePath;
 	private String fileSite;
@@ -28,37 +34,26 @@ public class Downloader {
 	private boolean hasInternet;
 	private boolean downloaded = false;
 	public Downloader(String fileName, String dateName, String fileSite, String dateSite, long chunkSize){
-		filePath = System.getProperty("user.dir") + File.separatorChar + fileName;
-		datePath = System.getProperty("user.dir") + File.separatorChar + dateName;
-		this.fileSite = fileSite;
-		this.dateSite = dateSite;
-		System.out.println("Checking if internet is available...");
-		try {
-			InetAddress gitHub = InetAddress.getByName(new URL("https://github.com/").getHost());
-			if(gitHub.isReachable(5000)){
-				hasInternet = true;
-				System.out.println("Internet connection available");
-			}
-			else
-				hasInternet = false;	
-		} catch (UnknownHostException e) {
-			hasInternet = false;
-		} catch (MalformedURLException e) {
-			e.printStackTrace();
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
+		constructor(fileName, dateName, fileSite, dateSite);
+//		sets chunkSize, which is the amount of data downloaded per round
 		this.chunkSize = chunkSize;
 		usingChunks = true;
 	}
 	public Downloader(String fileName, String dateName, String fileSite, String dateSite){
+		constructor(fileName, dateName, fileSite, dateSite);
+		usingChunks = false;
+	}
+	public void constructor(String fileName, String dateName, String fileSite, String dateSite){
+		setPingURL();
+		setUserDir();
 		filePath = System.getProperty("user.dir") + File.separatorChar + fileName;
 		datePath = System.getProperty("user.dir") + File.separatorChar + dateName;
 		this.fileSite = fileSite;
 		this.dateSite = dateSite;
+//		checks if internet is available by pinging the URL specified in settings.txt
 		System.out.println("Checking if internet is available...");
 		try {
-			InetAddress gitHub = InetAddress.getByName(new URL("https://github.com/").getHost());
+			InetAddress gitHub = InetAddress.getByName(new URL(pingURL).getHost());
 			if(gitHub.isReachable(5000)){
 				hasInternet = true;
 				System.out.println("Internet connection available");
@@ -68,39 +63,57 @@ public class Downloader {
 		} catch (UnknownHostException e) {
 			hasInternet = false;
 		} catch (MalformedURLException e) {
-			e.printStackTrace();
+			ErrorLogger.logError(e);
+			return;
 		} catch (IOException e) {
-			e.printStackTrace();
+			ErrorLogger.logError(e);
+			return;
 		}
-		usingChunks = false;
 	}
 	public void checkForUpdate(){
+		scanAndDeleteOldFiles("tempFile", filePath.substring(filePath.lastIndexOf('.'), filePath.length()));
 		if(hasInternet){
-			if(new File(filePath).exists()){
-				if(new File(datePath).exists()){
-					System.out.println("Checking date of update...");
-					try {
-						URL date = new URL(dateSite);
-						ReadableByteChannel rbc = Channels.newChannel(date.openStream());
-						FileOutputStream fos = new FileOutputStream(System.getProperty("user.dir") + File.separatorChar + "tempDate.txt");
-						fos.getChannel().transferFrom(rbc, 0, Long.MAX_VALUE);
-						rbc.close();
-						fos.close();
-					} catch (MalformedURLException e) {
-						e.printStackTrace();
-					} catch (FileNotFoundException e) {
-						e.printStackTrace();
-					} catch (IOException e) {
-						e.printStackTrace();
-					}
-					Calendar timeOfUpdate = fileToCalendar(System.getProperty("user.dir") + File.separatorChar + "tempDate.txt");
-					Calendar timeOnFile = fileToCalendar(datePath);
-					new File(System.getProperty("user.dir") + File.separatorChar + "tempDate.txt").delete();
-					if(timeOnFile.compareTo(timeOfUpdate) > 0)
-						System.out.println("No download necessary");
-					else
-						downloadFiles();
+			if(new File(filePath).exists() && new File(datePath).exists()){
+//				the date of update on the website copy will be compared to the local copy to see if an update should happen
+				System.out.println("Checking date of update...");
+				ReadableByteChannel rbc = null;
+				FileOutputStream fos = null;
+				try {
+					URL date = new URL(dateSite);
+					rbc = Channels.newChannel(date.openStream());
+					fos = new FileOutputStream(System.getProperty("user.dir") + File.separatorChar + "tempDate.txt");
+					fos.getChannel().transferFrom(rbc, 0, Long.MAX_VALUE);
+				} catch (MalformedURLException e) {
+					ErrorLogger.logError(e);
+					return;
+				} catch (FileNotFoundException e) {
+					ErrorLogger.logError(e);
+					return;
+				} catch (IOException e) {
+					ErrorLogger.logError(e);
+					return;
+				} finally {
+					if(rbc != null)
+						try {
+							rbc.close();
+						} catch (IOException e) {
+							ErrorLogger.logError(e);
+							return;
+						}
+					if(fos != null)
+						try {
+							fos.close();
+						} catch (IOException e) {
+							ErrorLogger.logError(e);
+							return;
+						}
 				}
+				Calendar timeOfUpdate = fileToCalendar(System.getProperty("user.dir") + File.separatorChar + "tempDate.txt");
+				Calendar timeOnFile = fileToCalendar(datePath);
+				if(!new File(System.getProperty("user.dir") + File.separatorChar + "tempDate.txt").delete())
+					return;
+				if(timeOnFile.compareTo(timeOfUpdate) > 0)
+					System.out.println("No download necessary");
 				else
 					downloadFiles();
 			}
@@ -113,39 +126,75 @@ public class Downloader {
 	private void downloadFiles(){
 		System.out.println("Download necessary");
 		System.out.println("Downloading...");
+		ReadableByteChannel rbc = null;
+		FileOutputStream fos = null;
 		try {
+//			creates file in temporary directory to be downloaded to, so that the main file isn't lost if the download is aborted
+			File temp = File.createTempFile("tempFile", filePath.substring(filePath.lastIndexOf('.'), filePath.length()));
+//			the file will be deleted upon exit unless the download is aborted mid-way
+			temp.deleteOnExit();
 			URL file = new URL(fileSite);
 			URLConnection urlconnection = file.openConnection();
 			long size = urlconnection.getContentLength();
-			ReadableByteChannel rbc = Channels.newChannel(file.openStream());
-			FileOutputStream fos = new FileOutputStream(filePath);
+			rbc = Channels.newChannel(file.openStream());
+			fos = new FileOutputStream(temp);
+			System.out.println(temp.getAbsolutePath());
 			long position = 0;
 			if(usingChunks)
 				while(position < size){
+//					downloads chunkSize bytes and increases the position on the download accordingly
 					position += fos.getChannel().transferFrom(rbc, position, chunkSize);
+//					sets progress to the nearest tenth of the amount of bytes downloaded out of the total
 					progress = Math.round((float)(100 * (float)position / (float)size) * (float)10) / (float)10;
 					System.out.println(progress + "% done");
 				}
 			else
+//				if not printing percentage, it will download the whole thing in one round
 				fos.getChannel().transferFrom(rbc, 0, size);
-			rbc.close();
-			fos.close();
+			File actual = new File(filePath);
+//			the file will now be copied to the desired directory
+			if(actual.exists())
+				if(!actual.delete())
+					return;
+			Path temporary = temp.toPath();
+			Path desired = actual.toPath();
+			Files.copy(temporary, desired, REPLACE_EXISTING);
 		} catch (MalformedURLException e) {
-			e.printStackTrace();
+			ErrorLogger.logError(e);
+			return;
 		} catch (IOException e) {
-			e.printStackTrace();
+			ErrorLogger.logError(e);
+			return;
+		} finally {
+			if(rbc != null)
+				try {
+					rbc.close();
+				} catch (IOException e) {
+					ErrorLogger.logError(e);
+					return;
+				}
+			if(fos != null)
+				try {
+					fos.close();
+				} catch (IOException e) {
+					ErrorLogger.logError(e);
+					return;
+				}
 		}
 		System.out.println("Success");
 		System.out.println("Making datestamp...");
+//		a datestamp of when the update took place will be made
 		File dateFile = new File(datePath);
 		try {
 			dateFile.createNewFile();
 		} catch (IOException e) {
-			e.printStackTrace();
+			ErrorLogger.logError(e);
+			return;
 		}
 		if(dateFile.exists()){
 			Calendar calendar = Calendar.getInstance();
 			TimeZone utc = TimeZone.getTimeZone("UTC");
+//			all time is converted to UTC so that time zones don't screw up updating
 			calendar.setTimeZone(utc);
 			String[] values = new String[6];
 			values[0] = String.valueOf(calendar.get(Calendar.YEAR));
@@ -154,35 +203,70 @@ public class Downloader {
 			values[3] = String.valueOf(calendar.get(Calendar.HOUR_OF_DAY));
 			values[4] = String.valueOf(calendar.get(Calendar.MINUTE));
 			values[5] = String.valueOf(calendar.get(Calendar.SECOND));
+			BufferedWriter bw = null;
 			try {
 				FileWriter dateWriter = new FileWriter(dateFile);
-				BufferedWriter bw = new BufferedWriter(dateWriter);
+				bw = new BufferedWriter(dateWriter);
 				for(int i = 0; i < values.length; i++){
 					bw.write(values[i]);
 					bw.newLine();
 				}
-				bw.close();
 			} catch (IOException e) {
-				e.printStackTrace();
+				ErrorLogger.logError(e);
+				return;
+			} finally {
+				if(bw != null)
+					try {
+						bw.close();
+					} catch (IOException e) {
+						ErrorLogger.logError(e);
+					}
 			}
 			System.out.println("Datestamp success");
 			downloaded = true;
 		}
 	}
 	private Calendar fileToCalendar(String path){
+//		converts a file in the date format used in this program (line 1-year, line 2-month, line 3-day, line 4-hour, line 5-minute, line 6-second)
 		Scanner scanner = null;
+		int[] values = new int[6];
 		try {
 			scanner = new Scanner(new File(path));
+			for(int i = 0; i < values.length; i++)
+				values[i] = Integer.parseInt(scanner.nextLine());
 		} catch (FileNotFoundException e) {
-			e.printStackTrace();
+			ErrorLogger.logError(e);
+			return null;
+		} finally {
+			if(scanner != null)
+				scanner.close();
 		}
-		int[] values = new int[6];
-		for(int i = 0; i < values.length; i++)
-			values[i] = Integer.parseInt(scanner.nextLine());
-		scanner.close();
 		Calendar calendar = Calendar.getInstance();
 		calendar.set(values[0], values[1], values[2], values[3], values[4], values[5]);
 		return calendar;
+	}
+	private void scanAndDeleteOldFiles(String name, String suffix){
+//		if the download is aborted, a temporary file will be left behind.  this method deletes all temporary files left behind in the past
+		DirectoryStream<Path> ds = null;
+		try {
+			ds = Files.newDirectoryStream(Paths.get(System.getProperty("java.io.tmpdir")), name + '*' + suffix);
+			for(Path file : ds){
+				if(file.toFile().delete())
+					System.out.println("Old file " + file.toFile().getAbsolutePath() + " deleted successfully.");
+				else
+					System.out.println("Old file " + file.toFile().getAbsolutePath() + " denied being deleted.  That evil file!");
+			}
+		} catch (IOException e) {
+			ErrorLogger.logError(e);
+			return;
+		} finally {
+			try {
+				ds.close();
+			} catch (IOException e) {
+				ErrorLogger.logError(e);
+				return;
+			}
+		}
 	}
 	public float getProgress(){
 		return progress;
